move LocationComponent to Locations namespace
