+++ app/components/locations/location_component.html.erb
